+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 - bug fixes
   - error popups raised by the main process can be closed
-  - sliders don't send same value over network when the value change is below the widget's precision 
+  - sliders don't send same value over network when the value change is below the widget's precision
+- widgets
+  - **new** (piano) keyboard widget
 
 ## 0.11.6